Add TLS specific TrustAllCertificates methods (#120)
* Add TLS specific TrustAllCertificates methods

For Linux and OSX, nothing really changed. For windows, it will now allow a trust all certificates to work without forcing it to use LDAPS.

* Redo the methods for allowing TLSSTART without using SSLTLS

This grew to be a little more complicated than I had expected. Because the native methods are overrides of a central abstract method, all off them needed to have the optional parameter added, even though it isn't used in the Linux and Osx methods. However, changing this did allow me to remove the extra methods I had added earlier.